This is the R code for performing the clustering analysis of the ZSD data
We first load the data, then remove the columns with ID and PBD groups, we then go on to calculate the silhouette width to see the optimal number of clusters, then perform the PAM clustering and finally using tsne to plot the clusters on a 2-dimensional plot.

We then go on to compute a density graph (only used as visual representation) and bar graph for each variable.
As well as also performing a computation of each variable's p-value to see its statistical significance in the separation of the clusters.